requirements.txt
azure-cognitiveservices-speech==1.41.1
Django==4.2.17
Metaphone==0.6
numpy==1.26.4
python-dotenv==1.0.1
RapidFuzz==3.10.1
regex==2024.7.24
spacy
torch==2.4.1




